add zoom buttons actions
