+++ frontend/src/components/MainStage.tsx
@@ -6,6 +6,7 @@ import { LinkI } from "../models/link.model";
 import styles from "./MainStage.module.css";
 import SmallMenu from "./SmallMenu";
 import ZoomButtons from "./ZoomButtons";
+import { handleWheel } from "./../functions/mainStageActions";
 
 interface MainStageProps {
   blocks: BlockI[];
@@ -77,37 +78,6 @@ const MainStage = ({
     e.evt.preventDefault();
   }
 
-  const handleWheel = (e: any) => {
-    e.evt.preventDefault();
-
-    const stage: any = stageRef.current;
-    if (!stage) return;
-    const oldScale = stage.scaleX();
-    const pointer = stage.getPointerPosition();
-
-    const mousePointTo = {
-      x: (pointer.x - stage.x()) / oldScale,
-      y: (pointer.y - stage.y()) / oldScale,
-    };
-
-    let direction = e.evt.deltaY > 0 ? 1 : -1;
-
-    if (e.evt.ctrlKey) {
-      direction = -direction;
-    }
-
-    const scaleBy = 1.01;
-    const newScale = direction > 0 ? oldScale * scaleBy : oldScale / scaleBy;
-
-    const newPos = {
-      x: pointer.x - mousePointTo.x * newScale,
-      y: pointer.y - mousePointTo.y * newScale,
-    };
-
-    stage.scale({ x: newScale, y: newScale });
-    stage.position(newPos);
-  };
-
   return (
     <span>
       <Stage
@@ -122,7 +92,7 @@ const MainStage = ({
         }}
         className={styles.mainStage}
         ref={stageRef}
-        onWheel={handleWheel}
+        onWheel={(e) => handleWheel(e, stageRef)}
       >
         <FlowLayer
           blocks={blocks}
@@ -142,7 +112,7 @@ const MainStage = ({
         position_y={lastClickPosition.y}
         setIsBlockModalCreateVisible={setIsBlockModalCreateVisible}
       />
-      <ZoomButtons></ZoomButtons>
+      <ZoomButtons stage={stageRef.current}></ZoomButtons>
     </span>
   );
 };
