+++ frontend/src/components/ZoomButtons.tsx
@@ -7,18 +7,29 @@ import {
   faMagnifyingGlassPlus,
   faQuestion,
 } from "@fortawesome/free-solid-svg-icons";
+import zoomStage from "../functions/mainStageActions";
 
-interface ZoomButtonsProps {}
+interface ZoomButtonsProps {
+  stage: any;
+}
 
-const ZoomButtons = () => {
+const ZoomButtons = ({ stage }: ZoomButtonsProps) => {
   return (
     <div className={styles.zoomButtons}>
       <Card className={styles.buttonsCard}>
-        <Button variant="secondary" className={styles.topButton}>
+        <Button
+          variant="secondary"
+          className={styles.topButton}
+          onClick={() => zoomStage(1, 20, stage)}
+        >
           {" "}
           <FontAwesomeIcon size="xs" icon={faMagnifyingGlassPlus} />
         </Button>
-        <Button variant="secondary" className={styles.bottomButton}>
+        <Button
+          variant="secondary"
+          className={styles.bottomButton}
+          onClick={() => zoomStage(-1, 20, stage)}
+        >
           {" "}
           <FontAwesomeIcon size="xs" icon={faMagnifyingGlassMinus} />
         </Button>
